refactore: date util
